+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - [BREAKING] Update ArnaBody
 - Update ArnaHeaderBar
 - Update ArnaScaffold
+- [BREAKING] Replace ArnaBottomBar with ArnaNavigationBar
+- [BREAKING] Replace ArnaBottomBarItem with ArnaNavigationBarItem
 - Add test for ArnaApp
 - Add test for ArnaScaffold
 
@@ -48,8 +50,8 @@
 - Add ArnaHeaderBarItem
 - Add OverflowHandler
 - [Example] Update Home
-- Add ArnaSlideTransition 
-- Add ArnaFadeTransition  
+- Add ArnaSlideTransition
+- Add ArnaFadeTransition
 - Update ArnaDivider
 - Update ArnaTabView
 - Update ArnaTabItem